feat: add account storage and authentication
This commit adds support for configuring account storage and
authentication in Kmail.

An accounts file can now be specified in the config, which stores
email-password pairs. Passwords can be stored as plaintext or argon2
hashes. Accounts are used for SMTP and IMAP/POP3 authentication.

The commit also adds the ability to append messages via IMAP.